+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/mattlewis92/angularx-flatpickr/compare/v1.0.0...v2.0.0) (2017-08-11)
+
+
+### Bug Fixes
+
+* make the directive to work with reactive forms ([e619706](https://github.com/mattlewis92/angularx-flatpickr/commit/e619706)), closes [#1](https://github.com/mattlewis92/angularx-flatpickr/issues/1)
+
+
+### Features
+
+* upgrade flatpickr to v3 ([ea73c66](https://github.com/mattlewis92/angularx-flatpickr/commit/ea73c66))
+
+
+### BREAKING CHANGES
+
+* flatpickr has been upgraded to v3. The `utc` was removed from this lib as it was
+removed from flatpickr
+* the forms module is now required to use this component
+
+
+
 <a name="1.0.0"></a>
 # 1.0.0 (2017-04-07)